requirements.txt
django==3.2.8
djangorestframework==3.12.4
django-rest-framework-simplejwt==5.0.0
